Refactoring: extract payment module (neon-evm #92)
