--- evm_loader/program/src/constant.rs
@@ -1,10 +1,0 @@
-pub const COLLATERAL_POOL: &str = "4sW3SZDJB7qXUyCYKA7pFL8eCTfm3REr8oSiKkww7MaT";
-pub const COLLATERAL_SEEDS: &[&str] = &[
-    "collateral_seed_0",
-    "collateral_seed_1",
-    "collateral_seed_2",
-    "collateral_seed_3",
-    "collateral_seed_4",
-];
-
-pub const PAYMENT_TO_COLLATERAL_POOL: u64 = 1000;
