+++ evm_loader/program/src/entrypoint.rs
@@ -6,7 +6,6 @@ use std::{alloc::Layout, mem::size_of, ptr::null_mut, usize};
 //use crate::{error::TokenError, processor::Processor};
 //use arrayref::{array_ref, array_refs, array_mut_ref, mut_array_refs};
 use std::convert::{TryFrom, TryInto};
-use std::str::FromStr;
 
 use evm::{
     ExitError, ExitFatal, ExitReason, ExitSucceed,
@@ -24,11 +23,11 @@ use crate::{
 //    bump_allocator::BumpAllocator,
 account_data::{Account, AccountData, Contract},
 account_storage::{ProgramAccountStorage, Sender},
-constant,
 error::EvmLoaderError,
 executor::Machine,
 executor_state::{ExecutorState, ExecutorSubstate},
 instruction::{EvmInstruction, on_event, on_return},
+payment,
 solana_backend::{AccountStorage, SolanaBackend},
 solidity_account::SolidityAccount,
 storage_account::StorageAccount,
@@ -231,8 +230,8 @@ fn process_instruction<'a>(
             let sysvar_info = next_account_info(account_info_iter)?;
             let operator_sol_info = next_account_info(account_info_iter)?;
             let collateral_pool_sol_info = next_account_info(account_info_iter)?;
-            let user_eth_info = next_account_info(account_info_iter)?;
-            let operator_eth_info = next_account_info(account_info_iter)?;
+            let _user_eth_info = next_account_info(account_info_iter)?;
+            let _operator_eth_info = next_account_info(account_info_iter)?;
             let system_info = next_account_info(account_info_iter)?;
 
             let _program_info = next_account_info(account_info_iter)?;
@@ -247,14 +246,12 @@ fn process_instruction<'a>(
                 account_storage.get_caller_account().ok_or(ProgramError::InvalidArgument)?,
                 &H160::from_slice(from_addr), trx.nonce, &trx.chain_id)?;
 
-            check_collateral_account(program_id,
-                                     collateral_pool_sol_info,
-                                     collateral_pool_seed_index[0] as usize)?;
-            perform_payments(operator_sol_info,
-                             collateral_pool_sol_info,
-                             user_eth_info,
-                             operator_eth_info,
-                             system_info)?;
+            payment::check_collateral_account(program_id,
+                                              collateral_pool_sol_info,
+                                              collateral_pool_seed_index[0] as usize)?;
+            payment::operator(operator_sol_info,
+                              collateral_pool_sol_info,
+                              system_info)?;
 
             let trx_gas_limit = u64::try_from(trx.gas_limit).map_err(|_| ProgramError::InvalidInstructionData)?;
             do_call(program_id, &mut account_storage, &accounts[5..], trx.call_data, trx_gas_limit)
@@ -705,59 +702,6 @@ fn check_ethereum_authority<'a>(
                 chain_id, SolanaBackend::<ProgramAccountStorage>::chain_id());
         return Err(ProgramError::InvalidArgument);
     }
-
-    Ok(())
-}
-
-/// Checks collateral accounts for the Ethereum transaction execution.
-#[allow(clippy::unnecessary_wraps)]
-fn check_collateral_account(program_id: &Pubkey,
-                            collateral_pool_sol_info: &AccountInfo,
-                            collateral_pool_seed_index: usize) -> ProgramResult {
-    debug_print!("program_id {:?}", program_id);
-    debug_print!("collateral_pool_sol_info {:?}", collateral_pool_sol_info);
-    debug_print!("collateral_pool_seed_index {}", collateral_pool_seed_index);
-
-    let collateral_pool = Pubkey::from_str(constant::COLLATERAL_POOL)
-        .map_err(|e| {
-            debug_print!("Error key string '{}', {:?}", constant::COLLATERAL_POOL, e);
-            ProgramError::InvalidArgument
-        })?;
-
-    if collateral_pool_seed_index >= constant::COLLATERAL_SEEDS.len() {
-        debug_print!("Error: seed index {} out of range [0..{}]",
-            collateral_pool_seed_index,
-            constant::COLLATERAL_SEEDS.len() - 1);
-        return Err(ProgramError::InvalidInstructionData);
-    }
-
-    let seed = constant::COLLATERAL_SEEDS[collateral_pool_seed_index];
-    let pool_key = Pubkey::create_with_seed(&collateral_pool, seed, program_id)?;
-    if *collateral_pool_sol_info.key != pool_key {
-        debug_print!("Collateral pool key {}", *collateral_pool_sol_info.key);
-        debug_print!("Wrong seed pool key {}", pool_key);
-        return Err(ProgramError::InvalidArgument);
-    }
-
-    Ok(())
-}
-
-/// Makes payments for the Ethereum transaction execution.
-#[allow(clippy::unnecessary_wraps)]
-fn perform_payments<'a>(operator_sol_info: &'a AccountInfo<'a>,
-                        collateral_pool_sol_info: &'a AccountInfo<'a>,
-                        user_eth_info: &'a AccountInfo<'a>,
-                        operator_eth_info: &'a AccountInfo<'a>,
-                        system_info: &'a AccountInfo<'a>) -> ProgramResult {
-    let transfer = system_instruction::transfer(operator_sol_info.key,
-                                                collateral_pool_sol_info.key,
-                                                constant::PAYMENT_TO_COLLATERAL_POOL);
-    let accounts = [(*operator_sol_info).clone(),
-                    (*collateral_pool_sol_info).clone(),
-                    (*user_eth_info).clone(),
-                    (*operator_eth_info).clone(),
-                    (*system_info).clone()];
-    invoke(&transfer, &accounts)?;
 
     Ok(())
 }
